Moved "getDataListFieldValues" method
